requirements.txt
PySide6==6.7.3
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
tenacity==9.0.0
tqdm==4.66.5
